Restructured stored_infotype files
